web/KangParkir/requirements.txt
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.1
Deprecated==1.2.18
Flask==3.1.0
Flask-Limiter==3.10.1
gcloud==0.18.3
googleapis-common-protos==1.68.0
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jwcrypto==1.5.6
limits==4.0.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
oauth2client==4.1.3
ordered-set==4.1.0
packaging==24.2
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.21.0
Pygments==2.19.1
pyparsing==3.2.1
pypng==0.20220715.0
PyQRCode==1.2.1
Pyrebase4==4.8.0
python-jwt==4.1.0
pytz==2025.1
requests==2.32.3
requests-toolbelt==0.10.1
rich==13.9.4
rsa==4.9
six==1.17.0
typing_extensions==4.12.2
urllib3==1.26.20
Werkzeug==3.1.3
wrapt==1.17.2
